Apply consistent export structure for JS modules
