+++ packages/checkbox/index.scss
@@ -1,4 +1,4 @@
-@forward "./src/variables";
-@forward "./src/functions";
-@forward "./src/checkbox";
-@forward "./src/checkbox_size";
+@forward "./src/css/variables";
+@forward "./src/css/functions";
+@forward "./src/css/checkbox";
+@forward "./src/css/checkbox_size";
