+++ packages/checkbox/src/css/_checkbox.scss
