+++ packages/checkbox/src/css/_checkbox_size.scss
