+++ packages/checkbox/src/css/_variables.scss
